using cookie

+++ client/src/pages/Admin.jsx
@@ -4,6 +4,7 @@ import { IoLogOut } from "react-icons/io5";
 import { useEffect, useState, useRef } from 'react';
 import { BsTrash2 } from "react-icons/bs";
 import { CgDanger } from "react-icons/cg";
+import { toast } from 'react-toastify';
 
 
 const API_URL = import.meta.env.VITE_API_URL
@@ -14,27 +15,28 @@ export default function Admin() {
 	const [selectedUser, setSelectedUser] = useState(null)
 	const navigate = useNavigate()
 	const token = localStorage.getItem('token')
-	if (!token) {
-		return <Navigate to="/" />
-	}
+	// if (!token) {
+	// 	return <Navigate to="/" />
+	// }
 
-	let decode
-	try {
-		decode = jwtDecode(token)
-	} catch (err) {
-		console.error("Invalid token:", err)
-		return <Navigate to="/" />
-	}
+	// let decode
+	// try {
+	// 	decode = jwtDecode(token)
+	// } catch (err) {
+	// 	console.error("Invalid token:", err)
+	// 	return <Navigate to="/" />
+	// }
 
-	if (!decode || decode.role !== 'admin') {
-		return <Navigate to="/" />
-	}
+	// if (!decode || decode.role !== 'admin') {
+	// 	return <Navigate to="/" />
+	// }
 
 	const fetchUser = async () => {
 		const response = await fetch(`${API_URL}/admin/user`, {
-			headers: {
-				'Authorization': `Bearer ${token}`
-			}
+			credentials: 'include'
+			// headers: {
+			// 	'Authorization': `Bearer ${token}`
+			// }
 		})
 		if (!response.ok) {
 
@@ -49,10 +51,22 @@ export default function Admin() {
 		fetchUser()
 	}, [])
 
-	const logout = () => {
-		localStorage.clear()
-		navigate('/')
+	const handleLogout = async () => {
+		const response = await fetch(`${API_URL}/logout`, {
+			method: 'POST',
+			credentials: 'include' // ⬅️ important pour envoyer le cookie
+		});
+
+		if (response.ok) {
+			// Nettoyage éventuel de localStorage si tu y mets d'autres choses
+			localStorage.clear()
+			navigate('/') // ou redirection vers la page de login
+		} else {
+			const data = await response.json()
+			console.error('Erreur de déconnexion', data)
+		}
 	}
+
 
 	const confirmDelete = (id) => {
 		console.log(id)
@@ -61,18 +75,19 @@ export default function Admin() {
 		modal.current.showModal()
 	}
 
-	const handleDelete = async() => {
-		const response = await fetch(`${API_URL}/admin/user/${selectedUser}`,{
+	const handleDelete = async () => {
+		const response = await fetch(`${API_URL}/admin/user/${selectedUser}`, {
 			method: 'DELETE',
-			headers: {
-				'Authorization': `Bearer ${token}`
-			}
+			// headers: {
+			// 	'Authorization': `Bearer ${token}`
+			// }
+			credentials: 'include'
 		})
+		const data = await response.json()
 		if (!response.ok) {
-
+			toast.success(data.message)
 		} else {
-			const data = await response.json()
-			console.log(data)
+			toast.success(data.message)
 			fetchUser()
 		}
 	}
@@ -82,7 +97,7 @@ export default function Admin() {
 		<div className='p-5'>
 			<div className='flex justify-between items-center'>
 				<h1 className='text-3xl font-bold'>Admin Dashboard</h1>
-				<button className='btn btn-warning' onClick={logout}><IoLogOut size={24} /></button>
+				<button className='btn btn-warning' onClick={handleLogout}><IoLogOut size={24} /></button>
 			</div>
 			<table className='table'>
 				<thead>
@@ -113,9 +128,9 @@ export default function Admin() {
 					<div className="modal-action">
 						<form method="dialog">
 							{/* if there is a button in form, it will close the modal */}
-							<button className="btn" onClick={()=>document.getElementById('delete_confirm').close()}>Annuler</button>
-							<button 
-								className="btn btn-error" 
+							<button className="btn" onClick={() => document.getElementById('delete_confirm').close()}>Annuler</button>
+							<button
+								className="btn btn-error"
 								onClick={handleDelete}>Confirmer</button>
 						</form>
 					</div>
